Fixed Some Issues
Uses chmod 700 to set executable bit (Windows doesn't have chmod, and the executable bit may be needed. I will have to see if it works)
Added ARM binary - note that it appears to be dynamically linked. It may not always work.
This may fix some bugs, and may allow the image shrinking to work on non-Windows ARM devices. I'm planning to test the application shortly on a computer that is not mine - because it is likely I made some changes that the end user will not have.