Read Discription
Fixed a major issue when compressing electron applications, where the node API would return the files inside the asar - which didn't exist to the outside system. This vastly improves compression speeds and can greatly improve compression performance, by allowing compression of files and directories that end in .asar

Note that the fix is not perfect - if made asyncronus, it would introduce a race condition, and may still do so being synchronus.

An issue will need to be opened on the electron repository.

Added fs-xattr fallback (mostly for offline development rather than another purpose)